requirements.txt
appdirs==1.4.4
attrs==20.2.0
certifi==2020.6.20
click==7.1.2
distlib==0.3.1
dnspython==2.0.0
filelock==3.0.12
Flask==1.1.2
Flask-PyMongo==2.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
pipenv==2020.8.13
pymongo==3.11.0
pyrsistent==0.16.0
python-bsonjs==0.2.0
six==1.15.0
virtualenv==20.0.31
virtualenv-clone==0.5.4
Werkzeug==1.0.1
